+++ android/src/main/java/com/fluttercandies/flutter_ali_auth/FlutterAliAuthPlugin.java
@@ -39,31 +39,23 @@ public class FlutterAliAuthPlugin implements FlutterPlugin,
     ///
     /// This local reference serves to register the plugin with the Flutter Engine and unregister it
     /// when the Flutter Engine is detached from the Activity
-    private MethodChannel mChannel;
+
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
-
-        mChannel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "flutter_ali_auth");
-
+        MethodChannel methodChannel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "flutter_ali_auth");
         authClient = AuthClient.getInstance();
-
-//        authClient.setChannel(channel);
-
-//        EventChannel auth_event = new EventChannel(flutterPluginBinding.getBinaryMessenger(), "auth_event");
-
+        authClient.setChannel(methodChannel);
+        methodChannel.setMethodCallHandler(this);
         authClient.setFlutterPluginBinding(flutterPluginBinding);
-
-//        auth_event.setStreamHandler(this);
-
-        mChannel.setMethodCallHandler(this);
-
     }
 
 
     @Override
     public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
+        Log.i(TAG,"call.method:"+call.method);
         switch (call.method) {
+
             case "getPlatformVersion":
                 result.success("Android " + android.os.Build.VERSION.RELEASE);
                 break;
@@ -71,13 +63,13 @@ public class FlutterAliAuthPlugin implements FlutterPlugin,
                 getAliAuthVersion(result);
                 break;
             case "init":
-                authClient.initSdk(call.arguments, result, mChannel);
+                authClient.initSdk(call.arguments, result);
                 break;
             case "login":
-                authClient.getLoginToken(call.arguments, result, mChannel);
+                authClient.getLoginToken(call.arguments, result);
                 break;
             case "loginWithConfig":
-                authClient.getLoginTokenWithConfig(call.arguments, result, mChannel);
+                authClient.getLoginTokenWithConfig(call.arguments, result);
                 break;
             case "hideLoginLoading":
                 authClient.hideLoginLoading();
@@ -90,60 +82,6 @@ public class FlutterAliAuthPlugin implements FlutterPlugin,
             default:
                 result.notImplemented();
         }
-//        switch (call.method) {
-//            case "getPlatformVersion":
-//                result.success("Android " + android.os.Build.VERSION.RELEASE);
-//                return;
-//            case "getAliAuthVersion":
-//                getAliAuthVersion(result);
-//                return;
-//            case "cancelStream":
-//                if (authClient.getEventSink() != null) {
-//                    authClient.getEventSink().endOfStream();
-//                }
-//                result.success(null);
-//                return;
-//            case "hideLoginLoading":
-//                authClient.hideLoginLoading();
-//                result.success(null);
-//                return;
-//            case "quitLoginPage":
-//                authClient.quitLoginPage();
-//                result.success(null);
-//                return;
-//        }
-//        if (Objects.isNull(authClient.getEventSink())) {
-//            AuthResponseModel authResponseModel = AuthResponseModel.initFailed(failedListeningMsg);
-//            result.success(authResponseModel.toJson());
-//            return;
-//        }
-        ///初始化相关
-//        switch (call.method) {
-//            case "init":
-//                authClient.initSdk(call.arguments, result);
-//                result.success(null);
-//                break;
-//            case "checkEnv":
-//                authClient.checkEnv();
-//                result.success(null);
-//                break;
-//            case "accelerateLoginPage":
-//                authClient.accelerateLoginPage();
-//                result.success(null);
-//                break;
-//            case "login":
-//                authClient.setLoginTimeout(AppUtils.integerTryParser(call.arguments, 5000));
-//                authClient.getLoginToken();
-//                result.success(null);
-//                break;
-//            case "loginWithConfig":
-//                authClient.getLoginToken(call.arguments);
-//                result.success(null);
-//                break;
-//            default:
-//                result.notImplemented();
-//                break;
-//        }
     }
 
     private void getAliAuthVersion(@NonNull Result result) {
@@ -154,7 +92,7 @@ public class FlutterAliAuthPlugin implements FlutterPlugin,
     @Override
     public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
         authClient.setFlutterPluginBinding(null);
-        mChannel.setMethodCallHandler(null);
+        authClient.getChannel().setMethodCallHandler(null);
     }
 
     @Override
@@ -174,8 +112,10 @@ public class FlutterAliAuthPlugin implements FlutterPlugin,
     @Override
     public void onDetachedFromActivity() {
         authClient.setFlutterPluginBinding(null);
-        mChannel.setMethodCallHandler(null);
+        authClient.getChannel().setMethodCallHandler(null);
     }
+
+
 
 //    @Override
 //    public void onListen(Object arguments, EventChannel.EventSink events) {